% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_phylo.R
\name{abundance_phylo}
\alias{abundance_phylo}
\title{Calculates the total abundance descending from each branch of a tree}
\usage{
abundance_phylo(tree, abundance_data)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{abundance_data}{A dataframe containing node abundance data for tree;
the first column must contain the node labels and second column the corresponding 
node abundance data; f no abundance data is known, don't input a dataframe and
code will assign leaves to be equally abundant and internal nodes to have size
zero}
}
\value{
A library, key is number assigned to node in 'phylo' object
}
\description{
Calculates the total abundance descending from each branch of a tree
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"), 
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
abundance_phylo(tree, tree_abundance)

}
