% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_indices.R
\name{all_indices}
\alias{all_indices}
\title{Calculates all index values}
\usage{
all_indices(file, node_abundances = FALSE, normalised = TRUE)
}
\arguments{
\item{file}{A string or phylo object; String of file name containing tree or 
the tree itself either as a string or a phylo object}

\item{node_abundances}{A dataframe containing node abundance data for tree;
the first column must contain the node labels and second column the corresponding 
node abundance data; if no abundance data is known, don't provide an input and
code will assign leaves to be equally abundant and internal nodes to have size
zero}

\item{normalised}{A boolean; either TRUE if the normalised indices are desired or
FALSE if the non-normalised indices are}
}
\value{
A dictionary containing all index values (for the key, 
if index is 1DS key is D1S)
}
\description{
Calculates all index values
}
\examples{
tree <- '(a:1,(b:0.5,c:0.5)e:0.5)d:0;'
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"), 
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
all_indices(tree, tree_abundance)

}
