% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DJ_i.R
\name{calculate_DJ_i}
\alias{calculate_DJ_i}
\title{Calculates qD_i or 1J_i for node i}
\usage{
calculate_DJ_i(tree, node, abundances, index_letter, q, individual)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{node}{A number; corresponds to number assigned in 'phylo' object}

\item{abundances}{Dictionary containing node abundances; output of [abundance_phylo()]}

\item{index_letter}{A string; a string containing a single letter, either 
"D" or "J" to specify desired index, if calculating all indices input does 
not matter}

\item{q}{A number; either 0 or 1 specifying desired index; if calculating
J only q=1 will be returned, if calculating all indices input does not matter}

\item{individual}{A boolean; either TRUE if only one index is desired or
FALSE if all indices are desired}
}
\value{
Either; a single number for the specified index, or a vector of all
index values (ordered 1D, 1J, 0D)
}
\description{
Calculates qD_i or 1J_i for node i
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"),
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
abundances <- abundance_phylo(tree, tree_abundance)
calculate_DJ_i(tree, 4, abundances, "D", 1, TRUE)

}
