% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DJ_i_a.R
\name{calculate_DJ_i_a}
\alias{calculate_DJ_i_a}
\title{Calculates 0D, 1D or 1J or all}
\usage{
calculate_DJ_i_a(
  tree,
  node,
  abundances,
  curr_ancestor,
  h,
  l_i,
  index_letter,
  q,
  individual
)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{node}{A number; corresponds to number assigned in 'phylo' object}

\item{abundances}{Dictionary containing node abundances; output of [abundance_phylo()]}

\item{curr_ancestor}{A number; corresponds to ancestor number assigned 
in 'phylo' object}

\item{h}{A number; distance between node and ancestor}

\item{l_i}{A number; length of longest direct descendant branch of node}

\item{index_letter}{A string; a string containing a single letter, either 
"D" or "J" to specify desired index, if calculating all indices input does 
not matter}

\item{q}{A number; either 0 or 1 specifying desired index; if calculating
J only q=1 will be returned, if calculating all indices input does not matter}

\item{individual}{A boolean; either TRUE if only one index is desired or
FALSE if all indices are desired}
}
\value{
Either; a list of dataframes if individual is FALSE, each dataframe
contains an index value and the corresponding value of x, key is index name;
if individual is TRUE, returns just one of these dataframes for the specified
index
}
\description{
Calculates 0D, 1D or 1J or all for given node i, ancestor a
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"), 
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
abundances <- abundance_phylo(tree, tree_abundance)
h <- distance(tree, 4,4)
l_i <- max(tree$edge.length[which(tree$edge[,1] == 4)])
calculate_DJ_i_a(tree,4,abundances,4,h,l_i, "D", 1, TRUE)

}
