% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_integral.R
\name{calculate_integral}
\alias{calculate_integral}
\title{Calculates integral for each ancestor}
\usage{
calculate_integral(
  tree,
  node,
  curr_ancestor,
  S_i,
  abundances,
  index_letter,
  q,
  individual
)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{node}{A number; corresponds to number assigned in 'phylo' object}

\item{curr_ancestor}{A number; corresponds to ancestor number assigned 
in 'phylo' object}

\item{S_i}{A dataframe; dataframe output from [compute_T_i_S_i()]}

\item{abundances}{Dictionary containing node abundances; output of [abundance_phylo()]}

\item{index_letter}{A string; a string containing a single letter, either 
"D" or "J" to specify desired index, if calculating all indices input does 
not matter}

\item{q}{A number; either 0 or 1 specifying desired index; if calculating
J only q=1 will be returned, if calculating all indices input does not matter}

\item{individual}{A boolean; either TRUE if only one index is desired or
FALSE if all indices are desired}
}
\value{
Either; a single number for the desired index, or a list of all indices
}
\description{
Calculates integral for each ancestor
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"), 
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
abundances <- abundance_phylo(tree, tree_abundance)
S_i <- compute_T_i_S_i(tree, 4, abundances)[[2]]
calculate_integral(tree, 4, 4, S_i, abundances, "D", 1, FALSE)

}
