% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_S_i_a.R
\name{calculate_S_i_a}
\alias{calculate_S_i_a}
\title{Calculate S_i_a for a given node i and ancestor a}
\usage{
calculate_S_i_a(tree, node, abundances, curr_ancestor, h, l_i)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{node}{A number; corresponds to number assigned in 'phylo' object}

\item{abundances}{Dictionary containing node abundances; output of [abundance_phylo()]}

\item{curr_ancestor}{A number; corresponds to ancestor number assigned 
in 'phylo' object}

\item{h}{A number; distance between node and ancestor}

\item{l_i}{A number; length of longest direct descendant branch of node}
}
\value{
A list with two elements; first element is a dataframe containing
value of S_i_a and corresponding value of x; second element is a dictionary
where each element contains every branch abundance present in each region of
x
}
\description{
Calculate S_i_a for a given node i and ancestor a
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"), 
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
abundances <- abundance_phylo(tree, tree_abundance)
h <- distance(tree, 4,5)
l_i <- max(tree$edge.length[which(tree$edge[,1] == 5)])
calculate_S_i_a(tree,5,abundances,4,h,l_i)

}
