% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_S_i_a_star.R
\name{calculate_S_i_a_star}
\alias{calculate_S_i_a_star}
\title{Calculate S_i_a for tree transformed to a star tree}
\usage{
calculate_S_i_a_star(tree, node, abundance_data)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{node}{A number; corresponds to number assigned in 'phylo' object}

\item{abundance_data}{Dictionary containing node abundances; output of 
[abundance_phylo()]}
}
\value{
A list with two elements; first element is a dataframe containing
value of S_i_a and corresponding value of x; second element is a dictionary
where each element contains every branch abundance present in each region of
x
}
\description{
Transforms tree into a star tree by attaching all of the trees branches to 
the root node and then calculates S_i_a
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"), 
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
calculate_S_i_a_star(tree,4,tree_abundance)

}
