% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_T_i_S_i.R
\name{compute_T_i_S_i}
\alias{compute_T_i_S_i}
\title{Calculates T_i and S_i for a given node i}
\usage{
compute_T_i_S_i(tree, node, abundances)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{node}{A number; corresponds to number assigned in 'phylo' object}

\item{abundances}{Dictionary containing node abundances; output of [abundance_phylo()]}
}
\value{
A list with 2 elements; first element is a number corresponding to T_i:
second element is a dataframe containing value of S_i and corresponding value 
of x
}
\description{
Calculates T_i and S_i for a given node i
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
tree_abundance <- data.frame("Names" = c("e", "d", "b", "c", "a"), 
                             "Values" = c(0, 0, 1/3, 1/3, 1/3))
abundances <- abundance_phylo(tree, tree_abundance)
compute_T_i_S_i(tree, 4, abundances)

}
