% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescendantEdges.R
\name{DescendantEdges}
\alias{DescendantEdges}
\title{Identify descendant edges}
\usage{
DescendantEdges(parent, child, edge, nEdge = length(parent))
}
\arguments{
\item{parent}{Vector containing first column of edge matrix}

\item{child}{Vector containing second column of edge matrox}

\item{edge}{Integer specifying the number of the edge whose children are
required (see \code{\link[ape:nodelabels]{edgelabels}()}).}

\item{nEdge}{number of edges (calculated from `length(parent)` if not
supplied).}
}
\value{
`DescendantEdges()` returns a logical vector stating whether each
edge in turn is the specified edge or one of its descendants.
}
\description{
Quickly identify edges that are "descended" from edges in a tree.
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
desc <- DescendantEdges(tree$edge[, 1], tree$edge[, 2], edge = 1)
}
\concept{tree navigation}
