% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tree.R
\name{dfToIgraph}
\alias{dfToIgraph}
\title{convert a data frame tree to igraph}
\usage{
dfToIgraph(dft)
}
\arguments{
\item{dft}{data frame tree}
}
\value{
igraph object and node attribute data frame
}
\description{
convert a data frame tree to igraph
}
