% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tree.R
\name{dfToNewick}
\alias{dfToNewick}
\title{convert data frame tree to Newick}
\usage{
dfToNewick(dft)
}
\arguments{
\item{dft}{data frame tree}
}
\value{
Newick string and node attribute data frame
}
\description{
convert data frame tree to Newick
}
