% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Calculates the  distance between two nodes}
\usage{
distance(tree, top_node, bottom_node)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{top_node}{A number; corresponds to number assigned in 'phylo' object;
must be ancestor of bottom_node}

\item{bottom_node}{A number; corresponds to number assigned in 'phylo' object;
must be a descendant of top_node}
}
\value{
A number
}
\description{
Calculates the distance between two specified nodes in terms of given
branch lengths.
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
distance(tree,4,1)

}
