% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_linear_phylo.r
\name{form_linear_phylo}
\alias{form_linear_phylo}
\title{Reads in a linear tree and outputs it as a phylo object}
\usage{
form_linear_phylo(file)
}
\arguments{
\item{file}{A string; String of the filename or the tree itself}
}
\value{
A 'phylo' object
}
\description{
Reads in a linear tree in newick format, either as a 
string or from a file and outputs it as a phylo object. 
If no branch lengthd are given they are assigned to be
one.
}
\examples{
tree_string <- '(((1:1)4:1)3:1)2:;'
tree_phylo <- form_linear_phylo(tree_string)

}
