% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descendant_branches.R
\name{get_descendant_branches}
\alias{get_descendant_branches}
\title{Calculates all descendant branches of a node}
\usage{
get_descendant_branches(tree, node)
}
\arguments{
\item{tree}{A tree as a 'phylo' object}

\item{node}{A number of node corresponding to number assigned in 'phylo'
object}
}
\value{
A dataframe containing for every descendant branch, the start and 
end corresponding to numbers assigned in 'phylo' object, branch length and 
distance from node to end of descendant branch length
}
\description{
Calculates all descendant branches of a node and records distance from node or "x"
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
get_descendant_branches(tree,4)
 
}
