% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllAncestors.R
\name{getAllAncestors}
\alias{getAllAncestors}
\title{Calculates all ancestor nodes of a given node including itself}
\usage{
getAllAncestors(tree, node)
}
\arguments{
\item{tree}{A 'phylo' object}

\item{node}{A number; corresponds to number assigned in 'phylo' object}
}
\value{
Vector of ancestor node numbers corresponding to numbers 
assigned in 'phylo' object
}
\description{
Calculates all ancestor nodes of a given node including itself
}
\examples{
tree <- ape::read.tree(text='(a:1,(b:0.5,c:0.5)e:0.5)d:0;')
getAllAncestors(tree,1)

}
