% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_convert.R
\name{read_convert}
\alias{read_convert}
\title{Reads in tree and converts to 'phylo' object}
\usage{
read_convert(file)
}
\arguments{
\item{file}{A string or phylo object; String of file name containing tree or 
the tree itself either as a string or a phylo object}
}
\value{
A 'phylo' object
}
\description{
Reads in tree from a file or tree can be inputted directly (e.g. as string),
and converts to 'phylo' object. If no branch lengths are given, they are 
assigned to be one.
}
\examples{
tree <- '(a:1,(b:0.5,c:0.5)e:0.5)d:0;'
read_convert(tree)

}
